Ask if continue with self-extracted java for linux

+++ orchestrate-java.sh
@@ -3,35 +3,44 @@
 javaDir=$pfDir/java
 makeDir $javaDir
 
-currentDir=`pwd`
-rm -rf /tmp/*
-for javaBinPath in `ls -d $cloudDir/java/$system-new/*bin 2>/dev/null`; do
-	javaBin=`basename $javaBinPath`
-	destDir=`echo $javaBin | sed 's/\.bin//g' | sed 's/j2sdk/jdk/g' | sed 's/amd64/x64/g' | sed "s/-$system//g"`
-	cp $javaBinPath /tmp/$javaBin
-	chmod u+x /tmp/$javaBin
-	workingDir="/tmp/$destDir"
-	mkdir $workingDir
-  cd $workingDir
-	/tmp/$javaBin
-	javaFile=`find "$workingDir" -name "java" 2>/dev/null | head -n 1`
+if [ "$system" == "linux" ] && [ `askYN "Configure Java from bin" "n"` == "y" ]; then
 
-	if [ -f "$javaFile" ]; then
-		if [ -d "$javaDir/$destDir" ]; then
-			echo -e "${CYAN}Dir $destFolder exists - skipping${NC}"
-		else
-			dirName=`ls $workingDir`
-			mkdir $javaDir/$destDir
-			cp -r $workingDir/$dirName/* $javaDir/$destDir/
+	currentDir=`pwd`
+	rm -rf /tmp/*
+	for javaBinPath in `ls -d $cloudDir/java/$system-bin/*bin 2>/dev/null`; do
+
+		javaBin=`basename $javaBinPath`
+		proceed=`askYN "Configure $javaBin" "n"`
+
+		if [ "$proceed" == "n" ]; then
+			continue
 		fi
-	else
-			echo -e "${RED}$workingDir does not contain java${NC}"
-	fi
+
+		destDir=`echo $javaBin | sed 's/\.bin//g' | sed 's/j2sdk/jdk/g' | sed 's/amd64/x64/g' | sed "s/-$system//g"`
+		cp $javaBinPath /tmp/$javaBin
+		chmod u+x /tmp/$javaBin
+		workingDir="/tmp/$destDir"
+		mkdir $workingDir
+	  cd $workingDir
+		/tmp/$javaBin
+		javaFile=`find "$workingDir" -name "java" 2>/dev/null | head -n 1`
+
+		if [ -f "$javaFile" ]; then
+			if [ -d "$javaDir/$destDir" ]; then
+				echo -e "${CYAN}Dir $destFolder exists - skipping${NC}"
+			else
+				dirName=`ls $workingDir`
+				mkdir $javaDir/$destDir
+				cp -r $workingDir/$dirName/* $javaDir/$destDir/
+			fi
+		else
+				echo -e "${RED}$workingDir does not contain java${NC}"
+		fi
+	done
+	cd $currentDir
+fi
 
 
-
-done
-cd $currentDir
 
 
 # for javaTgz in `ls -d $cloudDir/java/$system/*.tar.gz 2>/dev/null`; do
